chore: Completed work on dao tests and continued work on project setup.
Some modifications to dao tests may still be needed, but test coverage is largely complete.  Project now starts in dev mode using browser-sync and jade templates have been coverted to html.

Issuses: AC-851

+++ server/api/tagger/models/tag-area-target.js
@@ -6,7 +6,7 @@ module.exports = function(sequelize, DataTypes) {
     sequelize.define('TagAreaTarget',
     {
       id: {
-        type: DataTypes.INTEGER(5),
+        type: DataTypes.INTEGER(4),
         primaryKey: true,
         autoIncrement: true
       }
